# Don't disable GC.